added default image error

+++ src/components/Account.jsx
@@ -136,6 +136,10 @@ const Account = ({ Loader, Error, ImageLoader }) => {
   // delete user profile image
   const deleteProfilePic = async (e) => {
     e.preventDefault();
+    if (data.photo === "default.jpg") {
+      toast.error("Default image cannot be deleted");
+      return;
+    }
     const imageRef = ref(storage, imageURL);
     deleteObject(imageRef).then(() => {
       console.log("image delete successfully!");
